backend/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==3.2
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==23.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djoser==2.2.2
drf-yasg==1.21.7
flake8==4.0.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.6.1
oauthlib==3.2.2
packaging==24.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.22
pyflakes==2.4.0
PyJWT==2.9.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
